# Improve performance using a string builder instead of a string