--- a/company_test.go
+++ b/company_test.go
@@ -12,11 +12,25 @@ func TestCompanyCatchPhrase(t *testing.T) {
 	Expect(t, 2, strings.Count(phrase, " ")) // 3 words
 }
 
+func BenchmarkCompanyCatchPhrase(b *testing.B) {
+	f := New().Company()
+	for i := 0; i < b.N; i++ {
+		_ = f.CatchPhrase()
+	}
+}
+
 func TestCompanyBS(t *testing.T) {
 	f := New().Company()
 	bs := f.BS()
 	Expect(t, true, len(bs) > 0)
 	Expect(t, 2, strings.Count(bs, " ")) // 3 words
+}
+
+func BenchmarkCompanyBS(b *testing.B) {
+	f := New().Company()
+	for i := 0; i < b.N; i++ {
+		_ = f.BS()
+	}
 }
 
 func TestCompanySuffix(t *testing.T) {