**Description**

Use string builders instead of a string in functions that uses for loops to concatenate data. Here is the comparison between the main branch and this one. Data generated by benchstat.

```
name                  old time/op  new time/op  delta
BinaryString-8        7.78µs ± 0%  3.61µs ± 0%   ~     (p=1.000 n=1+1)
CompanyCatchPhrase-8   311ns ± 0%   226ns ± 0%   ~     (p=1.000 n=1+1)
CompanyBS-8            315ns ± 0%   231ns ± 0%   ~     (p=1.000 n=1+1)
```

**Are you trying to fix an existing issue?**

Yes #101 

**Go Version**

```
$ go version
go version go1.11 darwin/amd64
```

**Go Tests**

```
$ go test
ok  	faker	6.307s
```